Refactor `getRepoPath()` to use id not uuid.
This means we can set the ID. For our own repos it'll be whatever we put
into the uuid field currently, but later, it could be different. Also,
we need to choose a location for the repo BEFORE we know it's UUID if
it's a repo we're importing from elsewhere.