Bugfix. Don't assume octet-stream.
This doesn't work in iOS when reading a text file. Unclear why, but it throws because it trims the wrong number of leading characters and so produces a string that's not base64 encoded.

The approach here of using a regex is probably not super efficient, this could be improved performance wise I guess.