Rework app loading. Still not perfect.
The redux store and root saga start immediately, they log immediately, then asynchronously the logging enabled / disabled state is ready from AsyncStorage, and pushed into a variable. So it doesn't catch the first round of errors. Can potentially be improved, but I'm not sure it's worth the effort right now.